requirements.txt
python-dotenv==1.1.1
SQLAlchemy==2.0.43
aiogram==3.22.0